Cleaned Code
