+++ core/const.go
@@ -1,11 +1,12 @@
 package core
 
 const (
-	Blue    = "\u001b[38;5;77m"
+	Blue    = "\u001b[38;5;42m"
 	Green   = "\u001b[38;5;45m"
 	Grey    = "\u001b[38;5;252m"
 	Red     = "\u001b[38;5;42m"
 	White   = "\u001b[38;5;255m"
 	Reset   = "\u001b[0m"
+	Reverse = "\u001b[7m"
 	Version = "2.0"
 )
