+++ core/file.go
@@ -2,89 +2,30 @@ package core
 
 import (
 	"bufio"
-	"fmt"
 	"io/ioutil"
 	"log"
-	"net/http"
 	"os"
 	"path"
 	"path/filepath"
 	"regexp"
 	"strings"
-	"sync"
 )
 
-func GetData(url string) []byte {
-	VPrint(fmt.Sprintf("Fetching: %s\n", url))
+func FileValues(file string, array *[]string) {
 
-	res, err := http.Get(url)
+	readFile, err := os.Open(file)
+
 	if err != nil {
-		log.Fatal(err)
+		log.Fatalln("Err: Opening File ", file)
 	}
 
-	data, _ := ioutil.ReadAll(res.Body)
+	defer readFile.Close()
 
-	res.Body.Close()
-	return data
-}
+	fileScanner := bufio.NewScanner(readFile)
 
-func CheckFileCache(url string) {
-	filename := filepath.Base(url)
-
-	err := os.MkdirAll(path.Join(home, ".cache", "cook"), os.ModePerm)
-	if err != nil {
-		log.Fatalln("Err: Making .cache folder in HOME/USERPROFILE ", err)
-	}
-
-	if _, err := os.Stat(path.Join(home, ".cache", "cook", filename)); err != nil {
-		AppendToFile(path.Join(home, ".cache", "cook", filename), GetData(url))
-	}
-}
-
-func UpdateCache() {
-	fmt.Println(Banner)
-
-	goaddresses := make(chan string)
-	var wg sync.WaitGroup
-
-	for i := 0; i < 20; i++ {
-		go func() {
-			for file := range goaddresses {
-				func(file string) {
-					defer wg.Done()
-					filename := filepath.Base(file)
-					filepath := path.Join(home, ".cache", "cook", filename)
-					os.Remove(filepath)
-					AppendToFile(filepath, GetData(file))
-				}(file)
-			}
-		}()
-	}
-
-	for _, files := range M["files"] {
-		// fmt.Println("\n" + Blue + key + Reset)
-
-		for _, file := range files {
-			if strings.HasPrefix(file, "http://") || strings.HasPrefix(file, "https://") {
-				wg.Add(1)
-				goaddresses <- file
-			}
-		}
-	}
-
-	wg.Wait()
-}
-
-func AppendToFile(filepath string, data []byte) {
-	f, err := os.OpenFile(filepath, os.O_APPEND|os.O_WRONLY|os.O_CREATE, 0644)
-	if err != nil {
-		panic(err)
-	}
-
-	defer f.Close()
-
-	if _, err = f.Write(data); err != nil {
-		panic(err)
+	for fileScanner.Scan() {
+		line := strings.TrimRight(fileScanner.Text(), "\r")
+		*array = append(*array, line)
 	}
 }
 
@@ -115,23 +56,5 @@ func FindRegex(files []string, expresssion string, array *[]string) {
 		for k := range e {
 			*array = append(*array, k)
 		}
-	}
-}
-
-func FileValues(file string, array *[]string) {
-
-	readFile, err := os.Open(file)
-
-	if err != nil {
-		log.Fatalln("Err: Opening File ", file)
-	}
-
-	defer readFile.Close()
-
-	fileScanner := bufio.NewScanner(readFile)
-
-	for fileScanner.Scan() {
-		line := strings.TrimRight(fileScanner.Text(), "\r")
-		*array = append(*array, line)
 	}
 }
